UserService/requirements.txt
prometheus_client
Django>=3.2
django-redis
djangorestframework
psycopg2-binary

